+++ src/api/insight.ts
@@ -11,14 +11,13 @@ import {
   InsightPutRequest,
   RecommendGetResponse,
 } from '@/types/insight';
-import { useQuery } from '@tanstack/react-query';
 import axios from 'axios';
 
 interface SummarizeInsightResponse {
   choices: { message: { content: string } }[];
 }
 
-const fetchSummary = async (link: string, folderList: string[]) => {
+export const fetchSummary = async (link: string, folderList: string[]) => {
   const apiKey = process.env.NEXT_PUBLIC_GPT_API_KEY;
   const config = {
     method: 'post',
@@ -44,29 +43,6 @@ const fetchSummary = async (link: string, folderList: string[]) => {
   const response = await axios.request(config);
 
   return response.data as SummarizeInsightResponse;
-};
-
-export const useGetSummary = (link: string, folderList: string[]) => {
-  const { isLoading, error, data } = useQuery({
-    queryKey: ['get-summary'],
-    queryFn: () => fetchSummary(link, folderList),
-    enabled: !!link,
-  });
-  const result = {
-    title: data?.choices?.[0]?.message.content
-      .split('요약:')[0]
-      .split('제목:')[1],
-    summary: data?.choices?.[0]?.message.content
-      .split('요약:')[1]
-      .split('키워드:')[0],
-    keywords: data?.choices?.[0]?.message.content
-      .split('키워드:')[1]
-      .split('폴더명:')[0],
-    folderName: data?.choices?.[0]?.message.content
-      .split('폴더명: ')[1]
-      .split(','),
-  };
-  return { isLoading, error, result };
 };
 
 // 공유 폴더 url 보기
